+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/sdkconfiguration.py
@@ -20,8 +20,8 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.1'
-    sdk_version: str = '0.1.1'
-    gen_version: str = '2.116.0'
+    sdk_version: str = '0.1.2'
+    gen_version: str = '2.118.1'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: