+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.1.1] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [PyPI v0.1.1] https://pypi.org/project/codat-sync-for-commerce-version-1/0.1.1 - previous-versions/sync-for-commerce-version-1
+
+## 2023-09-20 00:15:23
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.88.0 (2.118.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.2] previous-versions/sync-for-commerce-version-1
+### Releases
+- [PyPI v0.1.2] https://pypi.org/project/codat-sync-for-commerce-version-1/0.1.2 - previous-versions/sync-for-commerce-version-1